Merge --squash 'logistics' branch into 13.0
All commits from the 12.0 logistics branch are squashed.
We have too many conflicts using cherry-pick or rebase, and
merge just adds more complexity when most of the commits in
the logistics branch were only about updating data files,
back and forth on songs and update on submodule. We have some
code too, but part of it will be removed anyway. If we want
the details of the commits, we can look in the 'logistics' branch.

It will reduce the size of the diff as well, as large files were
modified often.