BSCOS-1330 Add product_status + sale_exception
This module compute a status on product based on some date field
displayed on the Sales tab.

The option to use the product-attribute/product_state module was
considered but because the status is computed based on today. It does
not work well with a stored fields and also means there is no need for
creation of state from the user.

Add one exception and change the stock computation

Fix some data (product.tag, product.link.type)
Remove all product tags but `Customized`
Set the replacement type as not symetric