Ignore warehouse "block steps" constraints on install/updates
When we generate a database from scratch (dev or travis), cosanum_base
is installed in "setup", then again in version 13.0.4.0.0, forced with
`-i cosanum_base`. On the second "install" of the module, the write of
delivery and reception steps is done again, but would be blocked by the
constraint. Ignore it, if a module changes this, it knows what it is
doing :)