Fix auto-invoicing of grouped sales order pickings
When we have several orders grouped in the same transfer, we have a
single procurement group.

The method _get_sales_order_to_invoice() has
a local override to use the field "sale_ids". But this field (in
stock_picking_group_by_partner_by_carrier), uses
"procurement_group.sale_id" to find the related orders. As a result,
only the first sale (the one in the procurement group) will be invoiced.

We should not change the computation of the field "sale_ids", because it
links all the transfers of the chain, not only the OUT, its purpose is
not invoicing.

The original implementation of _get_sales_order_to_invoice() in
"account_invoice_mode_at_shipping" uses as well
"procurement_group.sale_id" (so the local override was useless).
I modified the original method to use
`"move_lines.sale_line_id.order_id` because it ensures we find all the
related sales for invoicing.

BSCOS-1550