Fix test based on default product.pricelist
The tests fail because there is this in ir.property now

 id  |            name            | res_id | company_id |   value_reference
-----+----------------------------+--------+------------+---------------------
  51 | property_product_pricelist |        |          2 | product.pricelist,2
 141 | property_product_pricelist |        |          1 | product.pricelist,3

A default value for the property_product_list of res_partner which will be
picked up as default value.

The issue probably comes from this change
https://github.com/odoo/odoo/commit/4b088d5523ea69c813a94147ec0644cb2ce77631

Which force the else case to be executed in more cases and a default value for
property_product_pricelist has to be written in ir_property during setup.

So the mechanism from product_pricelist_default module has no effect anymore
if a default value record exists in ir.property.
Because the default value is picked up before
https://github.com/odoo/odoo/blob/99531f5dd8966fc4ad89d1e9c2886ad701527f3a/addons/product/models/product_pricelist.py#L320

As those records do not exist in production database, this is only a problem in the tests.

But should the implementations of both module be reconsidered ?