[IMP] reception screen: minor improvements
Goal is to properly route the goods at reception step and provide a way
on the reception screen to know where is stored the product.

- invert the `set_location` and `select_packaging` steps
- filter the list of available destinations based on the package storage type
- product info: display the qty available, outgoing qty and the most recent expiry date

We also set the 'Shelving Bin' and 'Kardex Buffer In' location storage types
on 'WAE/PUTFBG*' and 'WAE/PUTSHU*' locations.

The 'Kardex Buffer Out' storage type has been created manually on production,
so it has been added as data beside the 'Kardex Buffer In'.

BSCOS-1967