Extract messages for "Delivery Carrier" and "Checkout Packing"
This PR adds the option to choose predefined messages for the
field "Info for Delivery Carrier", that is set on the res.partner
and reused in other places. For this, the original text field
delivery_info has been transformed into a Many2one that
points to the new model res.partner.delivery.info. Thus, the field
in the res.partner has been renamed to delivery_info_id,
that has caused some renames in other files that used it. These
new messages are archived instead of deleted, to keep track of
old messages associated to existing (or archived) res.partners.

Likewise for the "Checkout Packing Info", also in the res.partner.
This latter field is inside the module shopfloor_packing_info
of OCA/wms (see the URL at the bottom).

The new predefined messages can be created directly in the partner
or under the menu Inventory > Settings.

To migrate the data, the existing data has been migrated by assuming
that the first line is the name while all the content is for the text
of the new model.

Related OCA PR: https://github.com/OCA/wms/pull/161

Jira Card: BSCOS-1968