Missing translations for delivery note and package barcode
Added some missing translations to the delivery note
(for the stock.picking) and the package barcode (for
the stock.quant.package). Some translations overwrite
those of the core.

Also, in the package barcode report, the end of life
date shown is now a date, not a date+time as before.

Jira Card: BSCOS-2145
OCA PR: https://github.com/OCA/stock-logistics-workflow/pull/784