BSCOS-2572: Fix loading of "My Profile" when the user has an appraisal manager.
In standard odoo, the employee appraisal manager is a many2many_tags
widget trying to load the color of the employee. However, that color
field is restricted to "HR Officer" and cannot be read causing a ACL
error preventing to open the "My Profile" view.

Altough this error is raised on cosanum project, it was not successfully
reproduced on runbot.