Integrate CosaLog parcels with Shipment advices
It is done by changing the implementation of `display_name` on the
packages to display the list of CosaLog parcels instead of the package
name only if applicable.

As a side-effect the shipment advice is now printed with the right
package names, but this also shows the CosaLog parcel names also on the
picking form (and all places where the package is displayed).

By convenience the package kanban view has also been updated to show the
`display_name` instead of the `name` (view used both by the picking
and shipment smart buttons to list processed packages).

BSCOS-2756