cosanum_stock_warehouse: fix upgrade
Because of Shopinvader email constraints, we were unable to install
'cosanum_stock_warehouse' as this module updates the partner related to
the Werrikon warehouse which share the same email with other partners
in the database.

By disabling the constraint in 'pre' and re-enabling it again in 'post'
the module can be installed without issue.